Establish basic client/server communication.
The client is a simple shell that can relay a number of commands to the
server.  The server can reply in kind with simple strings.  There are so
many weaknesses and holes in this setup that it isn't even worth listing
them.

Expect the next few commits to streamline the network communication
process.